pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edgegrid-forecast-engine"
version = "0.1.0"
description = "Predictive dispatch engine for EdgeGrid — demand forecasting, solar generation prediction, BESS optimization, and energy market intelligence for India's distribution grid."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Praveen Peddi", email = "peddipra@gmail.com" },
]

dependencies = [
    # Core data
    "pandas>=2.0",
    "numpy>=1.24",
    "openpyxl>=3.1",

    # ML & Forecasting
    "lightgbm>=4.0",
    "scikit-learn>=1.3",
    "prophet>=1.1",
    "statsmodels>=0.14",

    # Solar physics
    "pvlib>=0.10",

    # Optimization
    "scipy>=1.11",
    "pulp>=2.7",

    # API
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.0",

    # Utilities
    "httpx>=0.24",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.5",
    "jupyter>=1.0",
    "matplotlib>=3.7",
    "plotly>=5.15",
    "seaborn>=0.12",
]

[project.scripts]
edgegrid-forecast = "edgegrid_forecast.api.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/edgegrid_forecast"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
